# connect4-openapi-spec.yml
openapi: 3.0.0
info:
  title: Connect4 Game API
  version: 1.0.0
  description: API for playing Connect4 against a bot

paths:
  /game/new:
    get:
      summary: Start a new Connect4 game
      description: Initializes a new game and returns an empty board
      operationId: startNewGame
      responses:
        '200':
          description: Successful response with a new game board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'

  /game/dropDisc:
    post:
      summary: Make a move by dropping a disc in the Connect4 game
      description: Allows a player to make a move in the game
      operationId: dropDisc
      requestBody:
        description: Information required to make a move
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DropDisc'
      responses:
        '200':
          description: Successful response with the updated game board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'

components:
  schemas:
    GameState:
      type: object
      properties:
        board:
          type: array
          description: 2D array representing the game board, using emoji
          items:
            type: array
            items:
              type: number
          example: [
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0]
          ]
        gameId:
          type: string
          description: Unique identifier for the game
          example: "game-123456"
      required:
        - board
        - gameId

    DropDisc:
      type: object
      properties:
        gameId:
          type: string
          description: Unique identifier for the game
          example: "game-123456"
        board:
          type: array
          description: Current state of the game board
          items:
            type: array
            items:
              type: number
          example: [
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 0, 0, 0, 0]
          ]
        column:
          type: number
          description: Column where the player wants to drop their piece
          enum: [0, 1, 2, 3, 4, 5, 6]
      required:
        - gameId
        - board
        - column